+++ src/wurfapi/doxygen_parser.py
@@ -587,22 +587,7 @@ def parse(xml, parser, log, scope):
     result["detaileddescription"] = parser.parse_element(
         xml=xml.find("detaileddescription"))
     result["access"] = xml.attrib["prot"]
-
-    identifier = {}
-    # The identifier type can be as just text in the type
-    # tag or in a nested ref tag. We use the approach
-    # mentioned here to get it:
-    # https://lxml.de/1.3/tutorial.html#elements-contain-text
-    identifier["type"] = xml.find("type").xpath("string()")
-
-    ref = xml.find("type/ref")
-
-    if ref is not None:
-        identifier["link"] = ref.attrib["refid"]
-    else:
-        identifier["link"] = None
-
-    result["identifier"] = identifier
+    result["type"] = parser.parse_element(xml=xml.find('type'))
 
     # Construct the unique name
     unique_name = scope + '::' + result["name"] if scope else result["name"]
@@ -629,6 +614,39 @@ def parse(xml, parser):
             result['line-stop'] = line_stop
     else:
         result['line-start'] = int(xml.attrib["line"])
+
+    return result
+
+
+@DoxygenParser.register(tag="type")
+def parse(xml, parser):
+    """ Parses Doxygen type
+
+    :return: Type list
+    """
+
+    result = []
+
+    def append_text(content):
+        if not content or content.isspace():
+            return
+        else:
+            result.append(
+                {"value": content.strip(), "link": None})
+
+    append_text(xml.text)
+
+    for child in xml.getchildren():
+
+        if match(xml=child, tag="ref"):
+
+            result.append(
+                {"value": child.text.strip(), "link": child.attrib["refid"]})
+
+        append_text(child.tail)
+
+    append_text(xml.tail)
+
     return result
 
 
@@ -640,28 +658,17 @@ def parse(xml, parser, log, scope):
     """
 
     result = {}
+    result["kind"] = "function"
+    result["scope"] = scope
     result['location'] = parser.parse_element(xml=xml.find('location'))
+    result["name"] = xml.findtext("name")
 
-    # First we get the name and type of the parameters
+    # Get the name and type of the parameters
     parameters = []
     for param in xml.findall('param'):
 
         parameter = {}
-
-        # The parameter type can be as just text in the type
-        # tag or in a nested ref tag. We use the approach
-        # mentioned here to get it:
-        # https://lxml.de/1.3/tutorial.html#elements-contain-text
-        parameter["type"] = param.find("type").xpath("string()")
-
-        ref = param.find("type/ref")
-
-        if ref is not None:
-            parameter["link"] = ref.attrib["refid"]
-        else:
-            parameter["link"] = None
-
-        # parameter['type'] = param.findtext('type')
+        parameter["type"] = parser.parse_element(xml=param.find("type"))
         parameter['name'] = param.findtext('declname')
         parameter['description'] = ''
 
@@ -701,26 +708,9 @@ def parse(xml, parser, log, scope):
     else:
         return_description = []
 
-    result["kind"] = "function"
-    result["scope"] = scope
-    result["name"] = xml.findtext("name")
-
     return_info = {}
-
-    # The return type can be as just text in the type
-    # tag or in a nested ref tag. We use the approach
-    # mentioned here to get it:
-    # https://lxml.de/1.3/tutorial.html#elements-contain-text
-    return_info["type"] = xml.find("type").xpath("string()")
+    return_info["type"] = parser.parse_element(xml=xml.find("type"))
     return_info["description"] = return_description
-
-    ref = xml.find("type/ref")
-
-    if ref is not None:
-        return_info["link"] = ref.attrib["refid"]
-    else:
-        return_info["link"] = None
-
     result["return"] = return_info
 
     result["signature"] = result["name"] + xml.findtext("argsstring")
@@ -738,7 +728,7 @@ def parse(xml, parser, log, scope):
 
     # If we do not have a return type the function is either a constructor
     # or a destructor
-    if return_info["type"] == "":
+    if return_info["type"] == []:
         result["is_constructor"] = not result["name"].startswith("~")
         result["is_destructor"] = result["name"].startswith("~")
     else:
@@ -749,7 +739,11 @@ def parse(xml, parser, log, scope):
     unique_name = scope + '::' + result["name"] if scope else result["name"]
 
     unique_name += '('
-    types = [parameter['type'] for parameter in parameters]
+    types = []
+    for parameter in parameters:
+        for value in parameter['type']:
+            types.append(value['value'])
+
     unique_name += ','.join(types)
     unique_name += ')'
 
@@ -765,23 +759,34 @@ def parse(xml, parser, log, scope):
 
     return {unique_name: result}
 
+
 def parse_variable_type(variable_type):
-    """ Parses the variable type
+    """ Parses the variable type list
     :return: (variable name, is const, is constexpr)
     """
-    if variable_type.find('constexpr ') != -1:
-        variable_type = variable_type.replace('constexpr ', '')
-        constexpr = True
-    else:
-        constexpr = False
 
-    if variable_type.find('const ') != -1:
-        variable_type = variable_type.replace('const ', '')
-        const = True
-    else:
-        const = False
+    # To update these variables from the nested function. See:
+    # https://stackoverflow.com/a/27004558/1717320
+    vars = {'is_const': False, 'is_constexpr': False}
 
-    return variable_type, const, constexpr
+    def prune(item):
+
+        if "constexpr " in item["value"]:
+            item["value"] = item["value"].replace("constexpr ", "")
+            vars['is_constexpr'] = True
+
+        if "const " in item["value"]:
+            item["value"] = item["value"].replace("const ", "")
+            vars['is_const'] = True
+
+        if item["value"]:
+            return item
+        else:
+            return None
+
+    variable_type = [item for item in variable_type if prune(item)]
+
+    return variable_type, vars['is_const'], vars['is_constexpr']
 
 
 @DoxygenParser.register(tag="memberdef", attrib={"kind": "variable"})
@@ -811,25 +816,11 @@ def parse(xml, parser, log, scope):
         v = v[2:]
     result["value"] = v
 
-    variable_type_info = {}
-
-    # The variable type can be as just text in the type
-    # tag or in a nested ref tag. We use the approach
-    # mentioned here to get it:
-    # https://lxml.de/1.3/tutorial.html#elements-contain-text
-    variable_type = xml.find("type").xpath("string()")
+    variable_type = parser.parse_element(xml=xml.find("type"))
 
     # Extract const and constexpr info from the variable type
-    variable_type, const, constexpr = parse_variable_type(variable_type)
-
-    variable_type_info["type"] = variable_type
-    ref = xml.find("type/ref")
-    if ref is not None:
-        variable_type_info["link"] = ref.attrib["refid"]
-    else:
-        variable_type_info["link"] = None
-
-    result["variable_type"] = variable_type_info
+    result["type"], const, constexpr = parse_variable_type(
+        variable_type)
 
     result['is_const'] = const
     result['is_constexpr'] = constexpr