+++ test/data/cpp_coffee/src/coffee/coffee.h
@@ -110,14 +110,23 @@ public:
     ///
     static std::string version();
 
-    /// Get the water tank
-    water_tank tank();
+    /// Get the first water tank
+    const water_tank& tank() const;
+
+    /// Get the first water tank
+    water_tank& tank();
+
+    /// Get all water tanks
+    std::vector<water_tank> tanks();
 
     /// The number of cups brewed by this machine.
     uint32_t cups_brewed = 0;
 
     /// The number of cups brewed by all machines.
     static uint32_t total_cups_brewed;
+
+    /// @return the mug_size
+    mug_size get_mug_size() const;
 
 protected:
     /// @brief Set the heat.